+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Alpha.0.5.2
 - Use TLS environment variable to dynamically enable HTTPS method.
 - Use environment variable to set admin user.
+- Use XYCHAT environment variable to determine that the application is running for what (test, debug or release).
 
 
 # Alpha.0.4